+++ src/components/FeedCard.jsx
@@ -4,32 +4,10 @@ import Badge from "./Badge";
 import { BiLike } from "react-icons/bi";
 import { BiDislike } from "react-icons/bi";
 import { postLikeDislike } from "../services/postLikeDislike";
+import LikeDislike from "./LikeDislike";
 
 const FeedCard = ({ question, userName, profileImage }) => {
   const [done, setDone] = useState(false);
-  const [isLikeClicked, setIsLikeClicked] = useState(false);
-  const [isDislikeClicked, setIsDislikeClicked] = useState(false);
-  const [like, setLike] = useState(question.like);
-  const [dislike, setDislike] = useState(question.dislike);
-
-  const handleClick = async ({ type }) => {
-    if (type === "like" && isLikeClicked === false) {
-      setIsLikeClicked(!isLikeClicked);
-      const response = await postLikeDislike({ type, id: question.id });
-      setLike(response.like);
-    } else if (type === "dislike" && isDislikeClicked === false) {
-      setIsDislikeClicked(!isDislikeClicked);
-      const response = await postLikeDislike({
-        type,
-        id: question.id,
-      });
-      setDislike(response.dislike);
-    } else if (type === "like" && isLikeClicked === true) {
-      setIsLikeClicked(!isLikeClicked);
-    } else if (type === "dislike" && isDislikeClicked === true) {
-      setIsDislikeClicked(!isDislikeClicked);
-    }
-  };
 
   useEffect(() => {
     if (question.answer?.id) {
@@ -77,22 +55,7 @@ const FeedCard = ({ question, userName, profileImage }) => {
       <div className="divider"></div>
 
       {/* 좋아요 싫어요 */}
-      <div className="like-dislike">
-        <div className={isLikeClicked ? "like-clicked" : "like"}>
-          <BiLike
-            onClick={() => handleClick({ type: "like" })}
-            className="like-icon"
-          />
-          <span>{like}</span>
-        </div>
-        <div className={isDislikeClicked ? "dislike-clicked" : "dislike"}>
-          <BiDislike
-            onClick={() => handleClick({ type: "dislike" })}
-            className="dislike-icon"
-          />
-          <span>{dislike}</span>
-        </div>
-      </div>
+      <LikeDislike question={question}/>
     </Card>
   );
 };
